+++ example/lib/app/view_app.dart
@@ -1,16 +1,25 @@
 //
 export 'logging/view/log_page.dart';
+
 //
-export 'my_app.dart';
+export 'view/my_app.dart';
+
 //
 //export 'two_tabs_scaffold.dart';
 //
 export 'settings/_view_settings.dart';
+
+//
+export 'view/common/dev_settings.dart';
+
 //
 export 'view/common/event_handler_calls.dart';
+
 //
 export 'view/common/splash_screen.dart';
+
 //
 export 'view/multi_tabs/multi_tabs_scaffold.dart';
+
 //
 export 'view/object/_view_app_object.dart';