+++ example/lib/home/controller/_controller_home.dart
@@ -1,8 +1,0 @@
-/// The export file for the directory, home.
-export 'another_controller.dart';
-
-export 'controller.dart';
-
-export 'word_pair_timer.dart';
-
-export 'yet_another_controller.dart';