+++ example/lib/home/controller_home.dart
@@ -1,0 +1,8 @@
+/// The export file for the directory, home.
+export 'controller/another_controller.dart';
+
+export 'controller/controller.dart';
+
+export 'controller/word_pair_timer.dart';
+
+export 'controller/yet_another_controller.dart';