+++ example/lib/home/model/_model_home.dart
@@ -1,4 +1,0 @@
-//
-export 'data_source.dart';
-
-export 'wordpairs_model.dart';