+++ example/lib/home/model_home.dart
@@ -1,0 +1,4 @@
+//
+export 'model/data_source.dart';
+
+export 'model/wordpairs_model.dart';