+++ example/lib/home/view/_view_home.dart
@@ -1,5 +1,0 @@
-/// The export file for the directory, view.
-
-export 'page_01.dart';
-export 'page_02.dart';
-export 'page_03.dart';