+++ example/lib/home/view/page_01.dart
@@ -297,6 +297,14 @@ class Page1State extends StateX<Page1> with EventsStateMixin {
 
     return caught;
   }
+
+  /// Using your favorite IDE, place a breakpoint in deactivate()
+  @override
+  Widget build(BuildContext context) => super.build(context);
+
+  @override
+  // ignore: unnecessary_overrides
+  void setState(VoidCallback fn, {bool? log}) => super.setState(fn, log: log);
 }
 
 ///