+++ example/lib/home/view_home.dart
@@ -1,0 +1,5 @@
+/// The export file for the directory, view.
+
+export 'view/page_01.dart';
+export 'view/page_02.dart';
+export 'view/page_03.dart';