+++ example/lib/src/view.dart
@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// Flutter's Material Interface package
-
 export 'package:flutter/cupertino.dart' hide RefreshCallback;
 
 export 'package:flutter/foundation.dart';
@@ -14,6 +12,6 @@ export 'package:state_extended/state_extended.dart';
 
 export '/another_app/_view_another_app.dart';
 
-export '/app/_view_app.dart';
+export '/app/view_app.dart';
 
-export '/home/view/_view_home.dart';
+export '/home/view_home.dart';